requirements.txt
uvicorn
fastapi
sqlalchemy
pymysql
python-dotenv